# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Management API
  version: 1.0.0
  description: API for managing inventory, products, users, and other related entities.
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /products/{id}/stock:
    get:
      summary: Get product stock by ID
      description: Retrieve the current stock level for a specific product by ID
      tags:
        - Products
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Product ID
      responses:
        '200':
          description: Current stock level
          content:
            application/json:
              schema:
                type: object
                properties:
                  stock:
                    type: integer
                    example: 100
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /products/{productId}/stock/batches:
    post:
      summary: Add new stock with batch information
      description: Adds a new batch of stock for a specific product
      tags:
        - Stock
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: productId
          schema:
            type: integer
          required: true
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockInBatchRequest'
      responses:
        '201':
          description: Batch created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Batch'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /products/{productId}/stock:
    get:
      summary: Get current stock levels for a product
      description: Retrieves current stock levels and batch breakdown for a specific product
      tags:
        - Stock
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: productId
          schema:
            type: integer
          required: true
          description: Product ID
        - in: query
          name: locationId
          schema:
            type: integer
          required: false
          description: Optional. Filter by Location ID
      responses:
        '200':
          description: Product stock details
          content:
            application/json:
              schema:
                type: object
                properties:
                  productId:
                    type: integer
                    example: 1
                  currentQuantity:
                    type: integer
                    example: 150
                  batches:
                    type: array
                    items:
                      $ref: '#/components/schemas/Batch'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /products/{productId}/stock/adjustments:
    post:
      summary: Perform a manual stock adjustment
      description: Performs a manual stock adjustment (stock-in or stock-out) for a product
      tags:
        - Stock
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: productId
          schema:
            type: integer
          required: true
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
      responses:
        '201':
          description: Stock adjustment successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Stock adjustment successful
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StockInBatchRequest:
      type: object
      required:
        - quantity
        - batchNumber
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
          example: 100
        batchNumber:
          type: string
          example: BATCH-XYZ-2023
        expiryDate:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-31T23:59:59Z"
...
    StockAdjustmentRequest:
      type: object
      required:
        - type
        - quantity
        - reasonCode
      properties:
        type:
          type: string
          enum: [STOCK_IN, STOCK_OUT]
          example: STOCK_IN
        quantity:
          type: integer
          format: int32
          minimum: 1
          example: 50
        reasonCode:
          type: string
          example: DAMAGED_GOODS
        notes:
          type: string
          example: Items damaged during transit
    Batch:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          readOnly: true
          example: 1
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        DeletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        ProductID:
          type: integer
          format: uint
          example: 1
        LocationID:
          type: integer
          format: uint
          example: 1
        BatchNumber:
          type: string
          example: BATCH-XYZ-2023
        Quantity:
          type: integer
          example: 100
        ExpiryDate:
          type: string
          format: date-time
          nullable: true
          example: "2025-12-31T23:59:59Z"
...